# docs/api/consumer/api.yaml
openapi: 3.0.1
info:
  title: Plotsensor consumer API
  description: API for my data
  termsOfService: http://me.com/terms
  contact:
    url: www.me.com
    email: me@me.com
  version: 0.0.1
servers:
- url: https://websensor.io/api/consumer/v1
paths:
...
  /captures:
    get:
      tags:
      - Open
      summary: Get a list of captures for a tag
      operationId: CapturesGet
      parameters:
      - name: serial
        in: query
        description: Tag serial
        required: true
        schema:
          type: string
      - name: offset
        in: query
        description: Return samples starting from this index.
        schema:
          type: integer
      - name: limit
        in: query
        description: Limit the number of samples returned.
        schema:
          type: integer
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        404:
          description: Tag with serial not found.
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
      - Open
      summary: Create a capture
      operationId: CapturesPost
      requestBody:
        description: Capture creation fields
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/EncodedCapture'
        required: true
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. HMAC does not correspond to input data.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Tag not found
          content:
            application/json:
              schema:
                type: object
        409:
          description: Conflict. A capture with this HMAC already exists. Dead battery
            or replay attack.
          content:
            application/json:
              schema:
                type: object
      x-codegen-request-body-name: body
  /captures/{id}:
    get:
      tags:
      - Open
      summary: Get a capture by ID
      operationId: CaptureGet
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        404:
          description: Capture not found.
          content:
            application/json:
              schema:
                type: object
...
  /locations:
    get:
      tags:
      - Open
      summary: Get a list of locations for a tag ordered by most recent
      security:
        - oAuthSample:
            - read_pets
      operationId: LocationGetList
      parameters:
      - name: serial
        in: query
        description: Tag serial
        required: true
        schema:
          type: string
      - name: starttime
        in: query
        description: start timestamp as an ISO-8601 string.
        schema:
          type: string
          format: datetime
      - name: endtime
        in: query
        description: end timestamp as an ISO-8601 string.
        schema:
          type: string
          format: datetime
      responses:
        200:
          description: A list of locations ordered from newest to oldest.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        404:
          description: Location not found.
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
      - Capture by User Required
      summary: Add location information to a tag
      operationId: LocationPost
      requestBody:
        description: Location fields
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Location'
        required: true
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. The user has no scanned this tag.
          content:
            application/json:
              schema:
                type: object
        403:
          description: Not authorised. Invalid JWT.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Parent resource not found.
          content:
            application/json:
              schema:
                type: object
      security:
      - Bearer: []
      x-codegen-request-body-name: body
  /locations/{id}:
    get:
      tags:
      - Open
      summary: Get a list of locations for a tag ordered by most recent
      operationId: LocationGet
      parameters:
      - name: id
        in: path
        description: Location ID
        schema:
          type: integer
      responses:
        200:
          description: A list of locations ordered from newest to oldest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        404:
          description: Location not found.
          content:
            application/json:
              schema:
                type: object
    delete:
      tags:
      - Capture by User Required
      summary: Delete a location
      operationId: LocationDelete
      responses:
        204:
          description: Location deleted
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad input parameter
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. The user has no scanned this tag.
          content:
            application/json:
              schema:
                type: object
        403:
          description: Not authorised. Invalid JWT.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Location not found.
          content:
            application/json:
              schema:
                type: object
      security:
      - Bearer: []
    patch:
      tags:
      - Capture by User Required
      summary: Edit location information for a tag
      operationId: LocationPut
      requestBody:
        description: Location fields
        content:
          '*/*':
            schema:
              type: string
              description: description
        required: true
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. The user has no scanned this tag.
          content:
            application/json:
              schema:
                type: object
        403:
          description: Not authorised. Invalid JWT.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Parent resource not found.
          content:
            application/json:
              schema:
                type: object
      security:
      - Bearer: []
      x-codegen-request-body-name: body
...
  /me/captures:
    get:
      tags:
      - Access Token Required
      summary: Get a list of captures taken by the current user ordered by most recent
        first.
      operationId: MeCapturesGet
      parameters:
      - name: distinctOnTag
        in: query
        description: Return only the latest capture for each scanned tag.
        schema:
          type: boolean
      responses:
        200:
          description: A list of capture objects ordered from newest to oldest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capture'
        400:
          description: invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. HMAC does not correspond to input data or invalid
            JWT.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Tag not found.
          content:
            application/json:
              schema:
                type: object
      security:
      - Bearer: []
    post:
      tags:
      - Access Token Required
      summary: Create a capture for a user
      operationId: MeCapturesPost
      requestBody:
        description: Capture creation fields
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/EncodedCapture'
        required: true
      responses:
        200:
          description: A capture object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        400:
          description: Invalid input, object invalid
          content:
            application/json:
              schema:
                type: object
        401:
          description: Not authorised. HMAC does not correspond to input data.
          content:
            application/json:
              schema:
                type: object
        403:
          description: Not authorised. Invalid JWT.
          content:
            application/json:
              schema:
                type: object
        404:
          description: Parent tag or user not found.
          content:
            application/json:
              schema:
                type: object
        409:
          description: Conflict. A capture with this HMAC already exists. Dead battery
            or replay attack.
          content:
            application/json:
              schema:
                type: object
      security:
      - Bearer: []
      x-codegen-request-body-name: body
...
components:
  schemas:
...
    Capture:
      title: Capture
      type: object
      properties:
        batvoltagemv:
          type: integer
          example: 3000
        tagserial:
          type: string
          description: Serial of the tag that made the capture
          example: YWJjZGVM
        cursorpos:
          type: integer
          example: 20
        id:
          type: integer
          format: int32
          example: 232
        loopcount:
          type: integer
          description: Number of times the circularbuffer pointer has wrapped from
            the end to the beginning
          example: 2
        md5:
          type: string
          example: 2f324022a223
        status:
          $ref: '#/components/schemas/CaptureStatus'
        timeintmins:
          type: integer
          example: 12
        timestamp:
          type: string
          format: date-time
          example: 2019-01-15T13:09:52.456Z
        version:
          type: integer
          example: 1
...
    CaptureStatus:
      title: CaptureStatus
      type: object
      properties:
        brownout:
          type: boolean
          description: True if the cause of the most recentreset was a brown out condition.
          example: false
        clockfail:
          type: boolean
          description: True if the most recent reset was caused by an MCU clock failure.
          example: false
        id:
          type: integer
          description: CaptureStatus ID
          format: int32
          example: 1
        lpm5wakeup:
          type: boolean
          description: True if the MCU has woken up from LPM (Low Power Mode) x.5
          example: false
        misc:
          type: boolean
          description: True if a reset has occured for miscellaneous reasons
          example: false
        parent_capture:
          type: integer
          description: Parent Capture ID
          format: int32
          example: 232
        resetsalltime:
          type: integer
          example: 20
        supervisor:
          type: boolean
          description: True if the supply voltage supervisor caused the most recent
            reset
          example: false
        watchdog:
          type: boolean
          description: True if the MCU watchdog caused the most recent reset.
          example: false
    EncodedCapture:
      title: Base64 Encoded Capture
      type: object
      properties:
        circbufb64:
          type: string
          description: Circular buffer including samples encoded as base64.
          example: Add example here
        serial:
          type: string
          description: Serial of the tag that made the capture
          example: YWJjZGVM
        statusb64:
          type: string
          description: 12 character status string
          example: AAAAAAAA
        timeintb64:
          type: string
          description: Time interval in minutes in base64
          example: Awg=
        versionStr:
          type: string
          example: "0001"
    Location:
      title: Location
      type: object
      properties:
        capturesample_id:
          type: integer
          description: ID of the parent CaptureSample object
          format: int32
          example: 1
        description:
          type: string
          description: Description of where a tag was located at the time of a capturesample.
          example: Cupboard under the stairs
        id:
          type: integer
          description: Location ID
          format: int32
          example: 1
        timestamp:
          type: string
          description: Unix timestamp in seconds since January 1st 1970
          format: date-time
          example: 2019-01-15T13:09:52.456Z
...
  securitySchemes:
    oAuthSample:    # <---- arbitrary name
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)
      flows:
        implicit:   # <---- OAuth flow(authorizationCode, implicit, password or clientCredentials)
          authorizationUrl: https://api.example.com/oauth2/authorize
          scopes:
            read_pets: read your pets
            write_pets: modify pets in your account
    Bearer:
      type: apiKey
      name: Authorization
      in: header